some extra to the fmt

+++ fmt/display.h
@@ -148,9 +148,9 @@ namespace fmt {
 
         public:
             void *data;
-            size_t data_size;
+            unsigned long long data_size;
 
-            fmtout(void *data = nullptr, size_t data_size = 0): data(data), data_size(data_size) {}
+            fmtout(void *data = nullptr, unsigned long long data_size = 0): data(data), data_size(data_size) {}
             
             // Overloaded operator<< to handle output
             template<typename T>
